+mockData, +postDisplay
how the post is going to look and some mock data of the database
