+++ client/src/components/sass/posts.scss
@@ -1,0 +1,47 @@
+.posts {
+    display: flex;
+    flex-direction: column;
+    margin-left: 40%;
+    .post {
+        border: lightgrey solid 2px;
+        border-radius: 20px;
+        margin-bottom: 1rem;
+        padding: 1rem 4rem;
+        width: 60%;
+        .prof {
+            display: flex;
+            flex-direction: row;
+            justify-content: center;
+            margin-bottom: 1rem;
+            &:first-child {
+                align-items: center;
+                img {
+                    height: 125px;
+                    width: auto;
+                    border-radius: 50%;
+                }
+            }
+            &:last-child {
+                display: flex;
+                flex-direction: column;
+                justify-content: flex-start;
+            }
+        }
+        .tags {
+            display: flex;
+            flex-direction: row;
+            &>* {
+                border: lightgrey solid 2px;
+                border-radius: 10px;
+                padding: .2rem .4rem;
+                margin-right: .2rem;
+            }
+        }
+        img {
+            border-radius: 10px;
+            position: relative;
+            width: 70%;
+            max-height: 70%;
+        }
+    }
+}
